requirements.txt
ah-jiplum==0.0.5
ah-datadog==0.0.5
arrow==0.14.4
blinker==1.4
datadog==0.39.0
ddtrace==0.47.0
dnspython==1.16.0
editdistance==0.4
fastapi==0.65.2
gunicorn==20.0.4
httptools==0.1.1
httpx==0.14.3
locustio==0.12.2
Logentries==0.17
matplotlib==3.0.2
mysqlclient==1.4.6
numpy==1.17.3
pandas==0.24.2
prometheus_client==0.7.1
python-json-logger==2.0.1
python-whois==0.7.2
pytz==2021.1
requests==2.23.0
sqlalchemy==1.3.16
starlette==0.14.2
starlette-prometheus==0.7.0
statsmodels==0.11.1
uvicorn==0.13.2
uvloop==0.14.0
